Fix Node 20 E2E bundle parse (Fixes #2240)
